Moved GetLocalTimeStamp into a Dataflow utility file.
GetLocalTimeStamp can now be found and imported from
src/dataflow/utilities/time.ts
