+++ src/dataflow/components/tools/dataflow-tool.tsx
@@ -7,7 +7,8 @@ import { IDocumentProperties } from "../../../lib/db-types";
 import { DocumentContentModel } from "../../../models/document/document-content";
 import { ToolTileModelType } from "../../../models/tools/tool-tile";
 import { DataflowContentModelType } from "../../models/tools/dataflow/dataflow-content";
-import { DataflowProgram, GetLocalTimeStamp } from "../dataflow-program";
+import { DataflowProgram } from "../dataflow-program";
+import { GetLocalTimeStamp } from "../../utilities/time";
 import { cloneDeep } from "lodash";
 import { SizeMe, SizeMeProps } from "react-sizeme";
 import "./dataflow-tool.sass";
